chore :📦 HistoryListViewModel 파일 삭제
- HistoryListPresentable 프로토콜을 TravelItemViewModel 파일로 이동
- HistoryListViewModel 파일 삭제
